controller/watcher/requirements.txt
rpi.gpio
web3
Pillow
